server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = ""
authors = [
    {name = "Joe Cardoso",email = "jow13_122@yahoo.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "pymongo (>=4.12.1,<5.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "flask-cors (>=5.0.1,<6.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "boto3 (>=1.38.21,<2.0.0)"
]

[tool.poetry]
packages = [{include = "services", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
